requirements.txt
alabaster==0.7.13
attrs==22.2.0
Babel==2.11.0
beautifulsoup4==4.11.1
certifi==2022.12.7
charset-normalizer==3.0.1
click==8.1.3
clickclick==20.10.2
connexion==2.14.2
docutils==0.19
Flask==2.2.2
furo==2022.12.7
future-fstrings==1.2.0
h11==0.14.0
idna==3.4
imagesize==1.4.1
inflection==0.5.1
iniconfig==2.0.0
itsdangerous==2.1.2
Jinja2==3.1.2
joblib==1.2.0
jsonschema==4.17.3
lark==1.1.5
lxml==4.9.2
MarkupSafe==2.1.2
networkx==3.0
ordered-set==4.1.0
packaging==23.0
piacere-doml-synthesis==2023.1.3
pluggy==1.0.0
pyecore==0.13.0
Pygments==2.14.0
pyrsistent==0.19.3
pytest==7.2.1
pytz==2022.7.1
PyYAML==6.0
requests==2.28.2
RestrictedPython==6.0
snowballstemmer==2.2.0
soupsieve==2.3.2.post1
Sphinx==6.1.3
sphinx-basic-ng==1.0.0b1
sphinxcontrib-applehelp==1.0.4
sphinxcontrib-devhelp==1.0.2
sphinxcontrib-htmlhelp==2.0.0
sphinxcontrib-jsmath==1.0.1
sphinxcontrib-qthelp==1.0.3
sphinxcontrib-serializinghtml==1.1.5
swagger-ui-bundle==0.0.9
termcolor==2.2.0
urllib3==1.26.14
uvicorn==0.20.0
Werkzeug==2.2.2
z3-solver==4.11.2.0
tabulate==0.9.0
pytest-xdist==3.3.1